backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ── Validation & settings ──────────────────────────────────────────────────────
pydantic[email]>=2.6.0
pydantic-settings>=2.2.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# ── Auth ───────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0   # JWT encode/decode
bcrypt>=4.1.2                      # password hashing
google-auth>=2.29.0                # Google ID token verification (React Native flow)
requests>=2.31.0                   # required by google-auth transport
python-multipart>=0.0.9            # form data (file uploads)

# ── HTTP client (Daily.co API) ─────────────────────────────────────────────────
httpx>=0.27.0

# ── Image processing ──────────────────────────────────────────────────────────
Pillow>=10.0.0                     # resize / compress profile photos

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pytz>=2024.1

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0                      # also used as async test client
